requirements.txt
google-genai
python-dotenv
sentence-transformers
faiss-cpu
python-docx
numpy